crispy forms added

+++ blog/views.py
@@ -1,8 +1,9 @@
+from http.client import HTTPResponse
 from multiprocessing import context
 from tkinter import N
 from django.shortcuts import render, redirect, get_object_or_404
 from .forms import CommentForm, PostForm
-from .models import Post
+from .models import Like, Post
 from django.template.defaultfilters import slugify
 # Create your views here.
 
@@ -61,6 +62,8 @@ def post_update(request,slug):
     #Post.objects.get(slug=lerafsdf-dgdfg-fghfg-h )
     obj = get_object_or_404(Post, slug=slug) # bu slugtakileri bul bana  listele demek
     #if req pos ise yerine bunu kullandık
+    if request.user.id != obj.author.id:
+        return redirect("blog:list", slug=slug)
     form = PostForm(request.POST or None, request.FILES or None, instance=obj)
     if form.is_valid():
         form.save()
@@ -73,6 +76,8 @@ def post_update(request,slug):
 
 def post_delete(request,slug):
     obj = get_object_or_404(Post, slug=slug)
+    if request.user.id != obj.author.id:
+        return HTTPResponse('You are not authorized !')
     if request.method == "POST":
         obj.delete()
         return redirect("blog:list")
@@ -80,3 +85,16 @@ def post_delete(request,slug):
         "object": obj,
         }
     return render(request, "blog/post_delete.html", context)
+
+
+def like(request, slug):
+    if request.method == "POST":
+        obj = obj = get_object_or_404(Post, slug=slug)
+        # artık hangi posta like yapacağımı biliyorum
+        like_qs = Like.objects.filter(user=request.user, post=obj)
+        if like_qs.exists():
+            like_qs[0].delete()
+        else:
+            Like.objects.create(user=request.user, post=obj)
+        return redirect("blog:detail", slug=slug)
+
